requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
requests==2.31.0